+++ src/base/src/types/math.rs
@@ -1,5 +1,5 @@
-use crate::{prelude::*, types, validator};
-use snafu::Snafu;
+use crate::{types, validator};
+use mimic::orm::prelude::*;
 
 ///
 /// DecimalFormat