+++ CHANGELOG.md
@@ -5,8 +5,11 @@ All notable, and occasionally less notable changes to this project will be docum
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-
 ## Unreleased
+
+## [0.0.3]
+
+- vanity release because now apps compile if they have mimic as a dependency!  go me!
 
 ## [0.0.2]